+++ Sources/RichTextKit/Attributes/RichTextAttributes.swift
@@ -11,35 +11,5 @@ import Foundation
 /**
  This typealias represents a ``RichTextAttribute`` keyed and
  `Any` valued dictionary.
-
- The typealias also defines additional functionality as type
- extensions for the platform-specific types.
  */
 public typealias RichTextAttributes = [RichTextAttribute: Any]
-
-public extension RichTextAttributes {
-
-    /**
-     Whether or not the attributes has a strikethrough style.
-     */
-    var isStrikethrough: Bool {
-        get {
-            self[.strikethroughStyle] as? Int == 1
-        }
-        set {
-            self[.strikethroughStyle] = newValue ? 1 : 0
-        }
-    }
-
-    /**
-     Whether or not the attributes has an underline style.
-     */
-    var isUnderlined: Bool {
-        get {
-            self[.underlineStyle] as? Int == 1
-        }
-        set {
-            self[.underlineStyle] = newValue ? 1 : 0
-        }
-    }
-}